feat(53-01): extract clipboardActions + add uiStore contextMenu/pendingLabelFocus slices
- New src/lib/clipboardActions.ts with copySelection/pasteSelection/hasClipboardContent/PASTE_OFFSET
- shortcuts.ts now imports from clipboardActions (no duplicate logic)
- uiStore: contextMenu, openContextMenu, closeContextMenu, pendingLabelFocus, setPendingLabelFocus
- Export ContextMenuState + ContextMenuKind types from uiStore
- 5 vitest tests (1 clipboard + 4 auto-flip math) passing
